split decentralized_dns into content and core
