+++ browser/decentralized_dns/test/BUILD.gn
@@ -12,10 +12,11 @@ source_set("browser_tests") {
   if (!is_android) {
     defines = [ "HAS_OUT_OF_PROC_TEST_RUNNER" ]
 
-    sources = [ "//brave/browser/decentralized_dns/test/decentralized_dns_navigation_throttle_browsertest.cc" ]
+    sources = [ "decentralized_dns_navigation_throttle_browsertest.cc" ]
     deps = [
       "//base/test:test_support",
-      "//brave/components/decentralized_dns",
+      "//brave/components/decentralized_dns/content",
+      "//brave/components/decentralized_dns/core",
       "//chrome/browser:browser_process",
       "//chrome/test:test_support",
       "//chrome/test:test_support_ui",
@@ -30,9 +31,9 @@ source_set("browser_tests") {
 source_set("unit_tests") {
   testonly = true
   sources = [
-    "//brave/browser/decentralized_dns/test/decentralized_dns_navigation_throttle_unittest.cc",
-    "//brave/browser/decentralized_dns/test/utils_unittest.cc",
     "//brave/browser/net/decentralized_dns_network_delegate_helper_unittest.cc",
+    "decentralized_dns_navigation_throttle_unittest.cc",
+    "utils_unittest.cc",
   ]
 
   deps = [
@@ -44,7 +45,8 @@ source_set("unit_tests") {
     "//brave/components/brave_wallet/browser:utils",
     "//brave/components/brave_wallet/browser/test:test_support",
     "//brave/components/brave_wallet/common:mojom",
-    "//brave/components/decentralized_dns",
+    "//brave/components/decentralized_dns/content",
+    "//brave/components/decentralized_dns/core",
     "//brave/components/tor/buildflags",
     "//chrome/test:test_support",
     "//components/prefs",
