+++ browser/ipfs/test/BUILD.gn
@@ -25,7 +25,8 @@ source_set("unittests") {
     "//brave/browser/net",
     "//brave/browser/tor",
     "//brave/common",
-    "//brave/components/decentralized_dns",
+    "//brave/components/decentralized_dns/content",
+    "//brave/components/decentralized_dns/core",
     "//brave/components/ipfs",
     "//brave/components/tor/buildflags",
     "//chrome/common:channel_info",
