+++ browser/net/BUILD.gn
@@ -72,7 +72,7 @@ source_set("net") {
     "//brave/components/brave_wallet/browser",
     "//brave/components/brave_wallet/common",
     "//brave/components/brave_webtorrent/browser/buildflags",
-    "//brave/components/decentralized_dns",
+    "//brave/components/decentralized_dns/content",
     "//brave/components/ipfs/buildflags",
     "//brave/components/update_client:buildflags",
     "//brave/extensions:common",
