+++ components/brave_wallet/browser/BUILD.gn
@@ -148,7 +148,7 @@ static_library("browser") {
     "//brave/components/brave_wallet/common:mojom",
     "//brave/components/brave_wallet/common:solana_utils",
     "//brave/components/constants",
-    "//brave/components/decentralized_dns",
+    "//brave/components/decentralized_dns/core",
     "//brave/components/filecoin/rs:cxx",
     "//brave/components/ipfs",
     "//brave/components/json/rs:cxx",
