+++ components/decentralized_dns/content/BUILD.gn
@@ -3,22 +3,19 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-static_library("decentralized_dns") {
+static_library("content") {
   sources = [
-    "constants.h",
     "decentralized_dns_interstitial_controller_client.cc",
     "decentralized_dns_interstitial_controller_client.h",
     "decentralized_dns_navigation_throttle.cc",
     "decentralized_dns_navigation_throttle.h",
     "decentralized_dns_opt_in_page.cc",
     "decentralized_dns_opt_in_page.h",
-    "pref_names.h",
-    "utils.cc",
-    "utils.h",
   ]
 
   deps = [
     "//base",
+    "//brave/components/decentralized_dns/core",
     "//brave/components/l10n/common",
     "//brave/components/resources:static_resources",
     "//brave/components/resources:strings",
@@ -28,7 +25,6 @@ static_library("decentralized_dns") {
     "//components/user_prefs",
     "//content/public/browser",
     "//net",
-    "//ui/base",
     "//url",
   ]
 }
