+++ components/decentralized_dns/content/DEPS
@@ -1,0 +1,3 @@
+include_rules = [
+  "+content/public/browser",
+]
