+++ components/decentralized_dns/content/decentralized_dns_interstitial_controller_client.h
@@ -3,8 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_DECENTRALIZED_DNS_DECENTRALIZED_DNS_INTERSTITIAL_CONTROLLER_CLIENT_H_
-#define BRAVE_COMPONENTS_DECENTRALIZED_DNS_DECENTRALIZED_DNS_INTERSTITIAL_CONTROLLER_CLIENT_H_
+#ifndef BRAVE_COMPONENTS_DECENTRALIZED_DNS_CONTENT_DECENTRALIZED_DNS_INTERSTITIAL_CONTROLLER_CLIENT_H_
+#define BRAVE_COMPONENTS_DECENTRALIZED_DNS_CONTENT_DECENTRALIZED_DNS_INTERSTITIAL_CONTROLLER_CLIENT_H_
 
 #include <memory>
 #include <string>
@@ -58,4 +58,4 @@ class DecentralizedDnsInterstitialControllerClient
 
 }  // namespace decentralized_dns
 
-#endif  // BRAVE_COMPONENTS_DECENTRALIZED_DNS_DECENTRALIZED_DNS_INTERSTITIAL_CONTROLLER_CLIENT_H_
+#endif  // BRAVE_COMPONENTS_DECENTRALIZED_DNS_CONTENT_DECENTRALIZED_DNS_INTERSTITIAL_CONTROLLER_CLIENT_H_
