+++ components/decentralized_dns/content/decentralized_dns_navigation_throttle.h
@@ -3,8 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_DECENTRALIZED_DNS_DECENTRALIZED_DNS_NAVIGATION_THROTTLE_H_
-#define BRAVE_COMPONENTS_DECENTRALIZED_DNS_DECENTRALIZED_DNS_NAVIGATION_THROTTLE_H_
+#ifndef BRAVE_COMPONENTS_DECENTRALIZED_DNS_CONTENT_DECENTRALIZED_DNS_NAVIGATION_THROTTLE_H_
+#define BRAVE_COMPONENTS_DECENTRALIZED_DNS_CONTENT_DECENTRALIZED_DNS_NAVIGATION_THROTTLE_H_
 
 #include <memory>
 #include <string>
@@ -54,4 +54,4 @@ class DecentralizedDnsNavigationThrottle : public content::NavigationThrottle {
 
 }  // namespace decentralized_dns
 
-#endif  // BRAVE_COMPONENTS_DECENTRALIZED_DNS_DECENTRALIZED_DNS_NAVIGATION_THROTTLE_H_
+#endif  // BRAVE_COMPONENTS_DECENTRALIZED_DNS_CONTENT_DECENTRALIZED_DNS_NAVIGATION_THROTTLE_H_
