+++ components/decentralized_dns/content/decentralized_dns_opt_in_page.cc
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/decentralized_dns/decentralized_dns_opt_in_page.h"
+#include "brave/components/decentralized_dns/content/decentralized_dns_opt_in_page.h"
 
 #include <ostream>
 #include <utility>
@@ -14,8 +14,8 @@
 #include "base/strings/string_util.h"
 #include "base/strings/utf_string_conversions.h"
 #include "base/values.h"
-#include "brave/components/decentralized_dns/decentralized_dns_interstitial_controller_client.h"
-#include "brave/components/decentralized_dns/utils.h"
+#include "brave/components/decentralized_dns/content/decentralized_dns_interstitial_controller_client.h"
+#include "brave/components/decentralized_dns/core/utils.h"
 #include "brave/components/l10n/common/locale_util.h"
 #include "components/grit/brave_components_resources.h"
 #include "components/grit/brave_components_strings.h"
