+++ components/decentralized_dns/content/decentralized_dns_opt_in_page.h
@@ -3,8 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_DECENTRALIZED_DNS_DECENTRALIZED_DNS_OPT_IN_PAGE_H_
-#define BRAVE_COMPONENTS_DECENTRALIZED_DNS_DECENTRALIZED_DNS_OPT_IN_PAGE_H_
+#ifndef BRAVE_COMPONENTS_DECENTRALIZED_DNS_CONTENT_DECENTRALIZED_DNS_OPT_IN_PAGE_H_
+#define BRAVE_COMPONENTS_DECENTRALIZED_DNS_CONTENT_DECENTRALIZED_DNS_OPT_IN_PAGE_H_
 
 #include <memory>
 #include <string>
@@ -61,4 +61,4 @@ class DecentralizedDnsOptInPage
 
 }  // namespace decentralized_dns
 
-#endif  // BRAVE_COMPONENTS_DECENTRALIZED_DNS_DECENTRALIZED_DNS_OPT_IN_PAGE_H_
+#endif  // BRAVE_COMPONENTS_DECENTRALIZED_DNS_CONTENT_DECENTRALIZED_DNS_OPT_IN_PAGE_H_
