+++ components/decentralized_dns/content/resources/decentralized_dns_interstitial.css
@@ -9,6 +9,6 @@
 
 .decentralized_dns .icon {
   background-image: -webkit-image-set(
-      url(../../../../components/security_interstitials/core/browser/resources/images/1x/info.png) 1x,
-      url(../../../../components/security_interstitials/core/browser/resources/images/2x/info.png) 2x);
+      url(../../../../../components/security_interstitials/core/browser/resources/images/1x/info.png) 1x,
+      url(../../../../../components/security_interstitials/core/browser/resources/images/2x/info.png) 2x);
 }
