+++ components/decentralized_dns/content/resources/decentralized_dns_interstitial.html
@@ -3,12 +3,12 @@
   <meta charset="utf-8">
   <meta name="theme-color" content="#fff">
   <title>$i18n{tabTitle}</title>
-  <link rel="stylesheet" href="../../../../components/security_interstitials/core/common/resources/interstitial_core.css">
-  <link rel="stylesheet" href="../../../../components/security_interstitials/core/common/resources/interstitial_common.css">
+  <link rel="stylesheet" href="../../../../../components/security_interstitials/core/common/resources/interstitial_core.css">
+  <link rel="stylesheet" href="../../../../../components/security_interstitials/core/common/resources/interstitial_common.css">
   <link rel="stylesheet" href="decentralized_dns_interstitial.css">
-  <script src="../../../../ui/webui/resources/js/assert.js"></script>
-  <script src="../../../../ui/webui/resources/js/util.js"></script>
-  <script src="../../../../components/security_interstitials/core/common/resources/interstitial_common.js"></script>
+  <script src="../../../../../ui/webui/resources/js/assert.js"></script>
+  <script src="../../../../../ui/webui/resources/js/util.js"></script>
+  <script src="../../../../../components/security_interstitials/core/common/resources/interstitial_common.js"></script>
   <script src="decentralized_dns_interstitial.js"></script>
 </head>
 <body id="body">
