+++ components/decentralized_dns/content/resources/decentralized_dns_interstitial.js
