+++ components/decentralized_dns/core/BUILD.gn
@@ -1,0 +1,20 @@
+# Copyright (c) 2022 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+static_library("core") {
+  sources = [
+    "constants.h",
+    "pref_names.h",
+    "utils.cc",
+    "utils.h",
+  ]
+
+  deps = [
+    "//base",
+    "//components/prefs",
+    "//net",
+    "//url",
+  ]
+}
