+++ components/decentralized_dns/core/constants.h
@@ -3,8 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_DECENTRALIZED_DNS_CONSTANTS_H_
-#define BRAVE_COMPONENTS_DECENTRALIZED_DNS_CONSTANTS_H_
+#ifndef BRAVE_COMPONENTS_DECENTRALIZED_DNS_CORE_CONSTANTS_H_
+#define BRAVE_COMPONENTS_DECENTRALIZED_DNS_CORE_CONSTANTS_H_
 
 namespace decentralized_dns {
 
@@ -21,4 +21,4 @@ enum class EnsOffchainResolveMethod { kAsk, kDisabled, kEnabled };
 
 }  // namespace decentralized_dns
 
-#endif  // BRAVE_COMPONENTS_DECENTRALIZED_DNS_CONSTANTS_H_
+#endif  // BRAVE_COMPONENTS_DECENTRALIZED_DNS_CORE_CONSTANTS_H_
