+++ components/decentralized_dns/core/pref_names.h
@@ -3,8 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_DECENTRALIZED_DNS_PREF_NAMES_H_
-#define BRAVE_COMPONENTS_DECENTRALIZED_DNS_PREF_NAMES_H_
+#ifndef BRAVE_COMPONENTS_DECENTRALIZED_DNS_CORE_PREF_NAMES_H_
+#define BRAVE_COMPONENTS_DECENTRALIZED_DNS_CORE_PREF_NAMES_H_
 
 namespace decentralized_dns {
 
@@ -28,4 +28,4 @@ constexpr char kEnsOffchainResolveMethod[] =
 
 }  // namespace decentralized_dns
 
-#endif  // BRAVE_COMPONENTS_DECENTRALIZED_DNS_PREF_NAMES_H_
+#endif  // BRAVE_COMPONENTS_DECENTRALIZED_DNS_CORE_PREF_NAMES_H_
