+++ components/resources/decentralized_dns_resources.grdp
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <grit-part>
-  <include name="IDR_DECENTRALIZED_DNS_INTERSTITIAL_HTML" file="../decentralized_dns/resources/decentralized_dns_interstitial.html" flattenhtml="true" type="BINDATA" />
-  <include name="IDR_DECENTRALIZED_DNS_INTERSTITIAL_JS" file="../decentralized_dns/resources/decentralized_dns_interstitial.js" type="BINDATA" />
-  <include name="IDR_DECENTRALIZED_DNS_INTERSTITIAL_CSS" file="../decentralized_dns/resources/decentralized_dns_interstitial.css" type="BINDATA" />
+  <include name="IDR_DECENTRALIZED_DNS_INTERSTITIAL_HTML" file="../decentralized_dns/content/resources/decentralized_dns_interstitial.html" flattenhtml="true" type="BINDATA" />
+  <include name="IDR_DECENTRALIZED_DNS_INTERSTITIAL_JS" file="../decentralized_dns/content/resources/decentralized_dns_interstitial.js" type="BINDATA" />
+  <include name="IDR_DECENTRALIZED_DNS_INTERSTITIAL_CSS" file="../decentralized_dns/content/resources/decentralized_dns_interstitial.css" type="BINDATA" />
 </grit-part>
